pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ffmc"
version = "1.0.0"
requires-python = ">=3.10"
dependencies = [
    "psutil>=5.9.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "aiohttp>=3.9.0"
]

[project.scripts]
ffmc = "ffmc.cli:main"